Update installation instructions for Ruby 2.4.9
This is the Ruby version CONSUL uses now. Thanks to that, we no longer
need to install an old version of libssl-dev for Ubuntu 18.04.

We could probably simplify the documentation using:

```bash
rbenv install `cat .ruby_version`
```

However, this would require downloading consul before installing Ruby,
which could be harder to document.