Fix hidden active elements in admin menu
In the admin menu, some links take you to a section, and some links open
a submenu with more links.

When we disable the "multi-open" property of the menu and the active
element is a link which takes you to a section, Foundation will hide it
whenever we click a link which opens a submenu.

The easiest solution is to enable "multi-open" property.