Destroy ckeditor instances before rendering the new page
Now we are using Turbolinks 5.X we are initializing application JS on
turbolinks:load event which is fired up on first page visits and on subsequent
Turbolinks page loads. So with this version of turbolinks we need to destroy
existing ckeditor instances before rendering the new page to avoid problem
when using browser history back. Inspiration here [1].

A similar version of this patch [2] was applied to a previous Consul version
where we were using turbolinks-classic version (deprecrated).

  [1] https://github.com/galetahub/ckeditor/issues/575#issuecomment-241185136
  [2] https://github.com/consul/consul/pull/3998