Use hidden input to store serialized form
Data attributes dynamically added to a page are lost when using browser
history so when user navigates back or forward to a page with a form with
unsaved changes the form serialization was done again but this time with
modified and unpersisted values so application thinks form wasn't changed.
By using a hidden input to store form serialized data there is no data loss
when using browser history and we can keep warning the user about unsaved
changes between the content original loaded (persisted) and the restored
content after using history back.