Sticky and Turbolinks 5 workaround
Destroy sticky elements before leaving the page to remove the window
scroll listener as it will fail when moving to any other page.

As we have destroyed sticky elements of any page before leaving
it now we need to initialize "again" all sticky elements when
restoring pages from browser cache, but restored HTML of the sticky
elements is not exactly as it was at first page load, more precisely
`data-sticky` attribute is lost so we need to add it again to all sticky
elements before calling `$(document).foundation()` initialization.

As `data-sticky` attribute was lost we need to rely on other attribute
or CSS class to be able to find sticky elements even when they lost
the `data-sticky` attribute.

Other problem i found is the usage of `data-stick-on` property as its
defined at Foundation docs [1]. we have only two sticky elements:

1. Support box on proposal show page

We need this one to be sticky only for mobile devices (small breakpoint),
but this was not possible by using `data-stick-on="small"` attribute as
it is, this was applying to all screen sizes as the Foundation definition
specifies so i had to adapt the initialization code so we can choose more
explicitly which breakpoints we want Sticky to be applied.

Before the allowed values for data-stick-on attribute had to be only one of the following

`data-stick-on=[small | medium | large | xlarge | xxlarge]`

Now we can add all breakpoints we need separated with whitespaces. If no
value is provided the default value is "small medium large xlarge xxlarge",
so it will be applied to all screen sizes.

`data-stick-on=[small medium large xlarge xxlarge]`

2. Table of contents of legislation processes draft version show page

This case is different as we want to it to be a sticky element for all
screen size so here we can omit the data-stick-on so it will be applied
to all breakpoints by default.

[1] https://get.foundation/sites/docs/sticky.html#js-options