Do not use data attributes for latitude and longitude when possible
When using an editable map is better to load marker latitude and longitude
information from form fields so we can restore the latest position defined
by user when the page was restored from browser history.

To reproduce this behavior:
0. Undo this commit
1. Go to new proposal page
2. Place the proposal map marker
3. Go away to any other page
4. Restore new proposal page from browser history.

At this point you should not see the recently placed marker.