Add contact information to footer
Note we're not using the Rails phone_to_number helper because it doesn't
seem to allow us to use the format we'd like to use.

Also note we have to redefine the styles for medium screens because the
email address might be too long to be seen properly in this context.

Co-Authored-By: Senén Rodero Rodríguez <senenrodero@gmail.com>