Add SDG goal model
The tests are saving the record before checking for uniqueness because
Globalize will consider the record valid before trying to save it.
That's because the unique title validation rule is applied to the
Translation model and not the Goal model. We have to apply it to the
Translation model because its unique by locale, since the same goal can
have the same title on different languages.

I'm not adding the same uniqueness validation rule to the description in
order to simplify the code and because, who knows, there might be a
scenario where two goals share the same description in a certain
language.

We're using silly examples in the tests and starting the factory
sequence at 50 in order to avoid potencial conflicts with real goals,
since we might add them to the seeds and so they will already exist in
the database when tests are run.